BE/Feat/#44 : create report API

See merge request s12-blockchain-nft-sub1/S12P21D210!62

+++ Backend/src/main/java/ssafy/d210/backend/dto/request/report/ReportRequest.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class ReportRequest {
     private long lectureId;
-    private long reportType;
+    private int reportType;
     private String reportContent;
 }
